Add a tf with subprocess instead of stdin
It does some other things nicer, but the final solution should be between this
and the previous thing (tf_shush). That one is certainly a lot less complex, and
the "correct" handling of stdout/stderr might be come for free there. Or not,
IDK.

I don't remember which of them has better error handling suggestions. It seems
like the other does, but I really thought it was this one. Some comparisons to
be done.

I'm just afraid of losing this work (again), since I previously deleted the
whole (of both scripts?) by accident :(